Seperate MyStoreInfoList and MyStoreTopBar
