+++ business/src/main/java/in/koreatech/business/feature/store/MyStoreDetailScreen.kt
@@ -90,155 +90,10 @@ fun MyStoreDetailScreen(
                 style = TextStyle(color = Color.White, fontSize = 18.sp),
             )
         }
-        MyStoreScrollScreen(state)
-
-
+        MyStoreScrollScreen(state, listState, pagerState)
     }
 }
 
-@Composable
-private fun CollapsedTopBar(
-    modifier: Modifier = Modifier, isCollapsed: Boolean
-) {
-    val viewModel: MyStoreDetailViewModel = hiltViewModel()
-    val state = viewModel.collectAsState().value
-    val color: Color by animateColorAsState(
-        if (isCollapsed) Color.White
-        else Color.Transparent
-    )
-    Box(
-        modifier = modifier
-            .background(color)
-            .fillMaxWidth(),
-        contentAlignment = Alignment.BottomStart
-    ) {
-        AnimatedVisibility(
-            modifier = Modifier.fillMaxWidth(),
-            visible = isCollapsed
-        ) {
-            Column(modifier = Modifier.fillMaxWidth()) {
-                Row(
-                    modifier = Modifier
-                        .fillMaxWidth()
-                        .height(50.dp)
-                        .padding(horizontal = 10.dp),
-                    horizontalArrangement = Arrangement.SpaceBetween,
-                    verticalAlignment = Alignment.CenterVertically,
-                ) {
-                    Text(text = state.storeInfo?.name ?: stringResource(R.string.shop_name))
-                    IconButton(onClick = { /*TODO*/ }) {
-                        Image(
-                            painter = painterResource(id = R.drawable.ic_setting),
-                            contentDescription = stringResource(R.string.shop_management)
-                        )
-                    }
-
-                }
-                Divider(
-                    color = ColorTextField,
-                    modifier = Modifier
-                        .fillMaxWidth()
-                        .height(1.dp)
-                )
-            }
-        }
-    }
-}
-
-@Composable
-fun TopBar() {
-    val viewModel: MyStoreDetailViewModel = hiltViewModel()
-    val state = viewModel.collectAsState().value
-    Column(modifier = Modifier.padding(horizontal = 20.dp)) {
-        Image(
-            modifier = Modifier
-                .fillMaxWidth()
-                .height(255.dp)
-                .padding(bottom = 10.dp),
-            painter = state.storeInfo?.imageUrls?.get(0)
-                .let { painterResource(id = R.drawable.no_image) },
-            contentDescription = stringResource(R.string.shop_image),
-            contentScale = ContentScale.Crop,
-        )
-
-        Button(
-            modifier = Modifier
-                .background(Color.White)
-                .fillMaxWidth()
-                .height(40.dp),
-            onClick = {},
-            colors = ButtonDefaults.buttonColors(
-                backgroundColor = Color.White,
-                contentColor = ColorPrimary,
-            ),
-            shape = Shapes.medium,
-            border = BorderStroke(1.dp, ColorPrimary)
-
-        ) {
-            Row(verticalAlignment = Alignment.CenterVertically) {
-                Image(
-                    painter = painterResource(id = R.drawable.ic_setting),
-                    contentDescription = stringResource(R.string.shop_management),
-                )
-                Text(text = stringResource(R.string.shop_management))
-            }
-
-        }
-        Text(
-            modifier = Modifier.padding(vertical = 10.dp),
-            text = state.storeInfo?.name ?: stringResource(R.string.shop_name),
-            style = TextStyle(color = Color.Black, fontSize = 20.sp),
-            fontSize = 20.sp,
-            fontWeight = FontWeight(600),
-        )
-    }
-
-}
-
-fun LazyListScope.storeDetailInfo(
-    state: MyStoreDetailState
-) {
-    val infoList = listOf("전화번호", "운영시간", "주소정보", "배달요금", "기타정보")
-    val dataList =
-        listOf(
-            state.storeInfo?.phone ?: "",
-            if (state.storeInfo?.open?.openTime?.isNotEmpty() == true && state.storeInfo.open.closeTime.isNotEmpty())
-                generateOpenCloseTimeString(
-                    state.storeInfo.open.openTime,
-                    state.storeInfo.open.closeTime
-                )
-            else
-                "",
-            state.storeInfo?.address ?: "",
-            state.storeInfo?.deliveryPrice.toString() + "원",
-            state.storeInfo?.description ?: "",
-        )
-    infoList.forEach {
-        info(it, dataList[infoList.indexOf(it)])
-    }
-}
-
-fun LazyListScope.info(info: String, data: String?) {
-    item {
-        Column(
-            modifier = Modifier.padding(horizontal = 20.dp, vertical = 10.dp)
-        ) {
-            Row(
-                modifier = Modifier.fillMaxWidth(),
-                horizontalArrangement = Arrangement.SpaceBetween
-            ) {
-                Text(
-                    text = info,
-                    style = TextStyle(color = Color.Black, fontSize = 15.sp),
-                )
-                Text(
-                    text = data ?: "",
-                    style = TextStyle(color = Color.Black, fontSize = 15.sp),
-                )
-            }
-        }
-    }
-}
 
 @SuppressLint("UnusedMaterialScaffoldPaddingParameter")
 @OptIn(ExperimentalFoundationApi::class)
